requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
asgiref==3.7.2
    # via django
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
cachetools==4.1.0
    # via google-auth
certifi==2024.7.4
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.1.0
    # via requests
contourpy==1.1.0
    # via matplotlib
cycler==0.10.0
    # via matplotlib
dj-database-url==2.0.0
    # via -r requirements.in
django==4.2.15
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-cors-headers
    #   djangorestframework
    #   drf-spectacular
django-cors-headers==4.1.0
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.27.2
    # via -r requirements.in
fonttools==4.43.0
    # via matplotlib
google-api-core==1.21.0
    # via google-api-python-client
google-api-python-client==1.9.3
    # via -r requirements.in
google-auth==1.18.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.0.3
    # via google-api-python-client
googleapis-common-protos==1.52.0
    # via google-api-core
gunicorn==22.0.0
    # via -r requirements.in
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.7
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonref==1.1.0
    # via -r requirements.in
jsonschema==4.21.1
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
kiwisolver==1.2.0
    # via matplotlib
lxml==4.9.1
    # via -r requirements.in
matplotlib==3.7.1
    # via -r requirements.in
numpy==1.25.0
    # via
    #   contourpy
    #   matplotlib
orjson==3.9.15
    # via yapw
packaging==23.1
    # via
    #   gunicorn
    #   matplotlib
pika==1.3.2
    # via yapw
pillow==10.3.0
    # via
    #   -r requirements.in
    #   matplotlib
protobuf==4.21.6
    # via
    #   google-api-core
    #   googleapis-common-protos
psycopg2==2.9.6
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyparsing==3.1.0
    # via
    #   httplib2
    #   matplotlib
python-dateutil==2.8.0
    # via matplotlib
pytz==2019.1
    # via google-api-core
pyyaml==6.0
    # via drf-spectacular
referencing==0.34.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via google-api-core
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
rsa==4.7.2
    # via google-auth
sentry-sdk==2.8.0
    # via -r requirements.in
setproctitle==1.2.2
    # via gunicorn
setuptools==74.1.2
    # via
    #   google-api-core
    #   google-auth
shortuuid==1.0.1
    # via -r requirements.in
simplejson==3.16.0
    # via -r requirements.in
six==1.16.0
    # via
    #   cycler
    #   google-api-core
    #   google-api-python-client
    #   google-auth
    #   python-dateutil
sqlparse==0.5.0
    # via django
typing-extensions==4.7.1
    # via dj-database-url
uritemplate==3.0.1
    # via
    #   drf-spectacular
    #   google-api-python-client
urllib3==2.2.2
    # via
    #   requests
    #   sentry-sdk
yapw==0.1.4
    # via -r requirements.in
